new routes working: establishment, product, address, redemptionCode

+++ src/controllers/productController.js
@@ -1,11 +1,18 @@
-const productService = require("../services/productService")
+const productService = require("../services/productService");
+const establishmentService = require("../services/establishmentService");
+const userService = require("../services/userService");
 
 const getProductByIdController = async (req, res) => {
   const { productId } = req.params;
+  const userId = req.user.idUtilizador;
   try {
-    const product = await productService.getProductById(productId);
+    const user = await userService.getUser(userId);
+    const neighborhoodId = user.VizinhançaidVizinhança;
+    const establishments = await establishmentService.getEstablishments(neighborhoodId);
+    const establishmentIds = establishments.map((e) => e.idEstabelecimento);
+    const product = await productService.getProductById(productId, establishmentIds);
 
-    res.status(200).json({ message: "Product fetched successfully", product: product });
+    res.status(200).json({ message: "Products fetched successfully", product: product });
   } catch (error) {
     res.status(500).json({ error: "Product not found" });
   }
@@ -13,17 +20,44 @@ const getProductByIdController = async (req, res) => {
 
 const getProductByTypeController = async (req, res) => {
   const { typeId } = req.params;
+  const userId = req.user.idUtilizador;
   try {
-    const product = await productService.getProductByType(typeId);
-    res.status(200).json({ message: "Product fetched successfully", product: product });
+    const user = await userService.getUser(userId);
+    const neighborhoodId = user.VizinhançaidVizinhança;
+    const establishments = await establishmentService.getEstablishments(neighborhoodId);
+    const establishmentIds = establishments.map((e) => e.idEstabelecimento);
+    const products = await productService.getProductByType(typeId,establishmentIds);
+    res.status(200).json({ message: "Products fetched successfully", products: products });
   } catch (error) {
     res.status(500).json({ error: "Product not found" });
   }
 };
 
-const getProductByEstablishmentController = async (req, res) => {
+const getProductController = async (req, res) => {
+  const userId = req.user.idUtilizador;
+
   try {
-    const { establishmentId } = req.params;
+    const user = await userService.getUser(userId);
+    const neighborhoodId = user.VizinhançaidVizinhança;
+    const establishments = await establishmentService.getEstablishments(neighborhoodId);
+    const establishmentIds = establishments.map((e) => e.idEstabelecimento);
+    const result = await productService.getProduct(establishmentIds);
+
+    res.status(200).json({message: "Products fetched successfully", products: result});
+  } catch (error) {
+    console.error("Error in getProductController:", error);
+    res.status(500).json({ error: "Internal server error" });
+  }
+};
+
+//Admin
+const getProductByEstablishmentController = async (req, res) => {
+  const userType = req.user.idTipoUtilizador;
+  const { establishmentId } = req.params;
+  try {
+    if (userType !== 1) {
+      return res.status(403).json({ error: "Access denied. Admins only." });
+    }
     const product = await productService.getProductByEstablishment(establishmentId);
     res.status(200).json({ message: "Product fetched successfully", product: product });
   } catch (error) {
@@ -31,24 +65,23 @@ const getProductByEstablishmentController = async (req, res) => {
   }
 };
 
-const getProductController = async (req, res) => {
-    try {
-        const result = await productService.getProduct();
-        res.status(200).json({ message: "Products fetched successfully", product: result  });
-    } catch (error) {
-        res.status(500).json("Internal server error");
-    };
-};
-
 const createProductController = async (req, res) => {
   try {
-    const { nomeProduto, precoProduto, descricaoProduto, tipoProdutoidTipoProduto, estadoProdutoidEstadoProduto, EstabelecimentoidEstabelecimento } = req.body;
-    const imagemProduto = req.file ? req.file.filename : null;
-
-    if (!imagemProduto) {
-      return res.status(400).json({ error: "Image upload failed" });
+    const userType = req.user.idTipoUtilizador;
+    if (userType !== 1) {
+      return res.status(403).json({ error: "Access denied. Admins only." });
     }
-
+    const {
+      nomeProduto,
+      precoProduto,
+      descricaoProduto,
+      tipoProdutoidTipoProduto,
+      estadoProdutoidEstadoProduto,
+      EstabelecimentoidEstabelecimento,
+    } = req.body;
+    const imagemProduto = req.file ? req.file.filename : null;
+    
+    
     const productData = {
       nomeProduto,
       precoProduto,
@@ -56,35 +89,53 @@ const createProductController = async (req, res) => {
       imagemProduto,
       tipoProdutoidTipoProduto,
       estadoProdutoidEstadoProduto,
-      EstabelecimentoidEstabelecimento
+      EstabelecimentoidEstabelecimento,
     };
 
     const result = await productService.createProduct(productData);
-    res.status(201).json({ message: "Product added successfully", product: result });
+    res
+      .status(201)
+      .json({ message: "Product added successfully", product: result });
   } catch (error) {
     console.error(error);
     res.status(500).json({ error: "Internal server error" });
   }
 };
 
-  const updateProductController = async (req, res) => {
-      try {
-        const { productId } = req.params;
-        await productService.updateProduct(productId, req.body);
-        res.status(200).json({ message: "Product updated successfully"});
-      } catch (error) {
-        res.status(500).json({ error: "Internal server error" });
-      }
-    };
+const updateProductController = async (req, res) => {
+  const userType = req.user.idTipoUtilizador;
+  try {
+    if (userType !== 1) {
+      return res.status(403).json({ error: "Access denied. Admins only." });
+    }
+    const { productId } = req.params;
+    await productService.updateProduct(productId, req.body);
+    res.status(200).json({ message: "Product updated successfully" });
+  } catch (error) {
+    res.status(500).json({ error: "Internal server error" });
+  }
+};
 
-  const deleteProductController = async (req, res) => {
-      const { productId } = req.params;
-      try {
-        await productService.deleteProduct(productId);
-        res.status(200).json({message: "Product deleted successfully"});
-      } catch (error) {
-        res.status(500).json({ error: "Internal server error" });
-      }
-    };
+const deleteProductController = async (req, res) => {
+  const { productId } = req.params;
+  const userType = req.user.idTipoUtilizador;
+  try {
+    if (userType !== 1) {
+      return res.status(403).json({ error: "Access denied. Admins only." });
+    }
+    await productService.deleteProduct(productId);
+    res.status(200).json({ message: "Product deleted successfully" });
+  } catch (error) {
+    res.status(500).json({ error: "Internal server error" });
+  }
+};
 
-module.exports = { getProductByIdController, getProductByTypeController, getProductByEstablishmentController, getProductController, createProductController, updateProductController, deleteProductController };
+module.exports = {
+  getProductByIdController,
+  getProductByTypeController,
+  getProductController,
+  getProductByEstablishmentController,
+  createProductController,
+  updateProductController,
+  deleteProductController,
+};
